python/bot/requirements_bot.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
MetaTrader5>=5.0
